refactor(operator_commands_dashboard): move login handlers + LOGIN_HTML to auth module
Cohesive consolidation: login/login_page HTTP handlers and LOGIN_HTML
template now live in auth.rs alongside try_login/require_auth/init_login_code.

- routes.rs: 5632 -> 5558
- auth.rs: 257 -> 339

Refs #1266
